+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ### Enhancements:
 
+- feat(config): Support file/format for kv_store and secret_store in fastly.toml
+- feat(config): Support metadata for kv_store in fastly.toml
+
 ### Bug fixes:
 
 ### Dependencies: